move osf.metadata.gather.osf up
